requirements.txt
up42-blockutils
tensorflow>=2.3.*
keras
numpy
rasterio
scikit-image
imageio
pyproj
geojson
shapely
ciso8601
tqdm
pytest
pytest-pylint
pytest-cov
pytest-mypy
mock
black
coverage-badge
